+++ src/Interface/SpacecraftInOut/Ports/I2CPort.h
@@ -19,8 +19,8 @@ class I2CPort {
   unsigned char ReadRegister(const unsigned char i2c_addr, const unsigned char reg_addr);
 
   // OBC->Component Command emulation
-  int WriteCommand(const unsigned char i2c_addr, const unsigned char* tx_data, const unsigned int length);
-  int ReadCommand(const unsigned char i2c_addr, unsigned char* rx_data, const unsigned int length);
+  size_t WriteCommand(const unsigned char i2c_addr, const unsigned char* tx_data, const size_t length);
+  size_t ReadCommand(const unsigned char i2c_addr, unsigned char* rx_data, const size_t length);
 
  private:
   unsigned char max_register_number_ = 0xff;
